Correções e testes para SonarCloud

+++ src/main/java/br/edu/ifgoias/academico/repositories/CursoRepository.java
@@ -1,11 +1,9 @@
 package br.edu.ifgoias.academico.repositories;
 
+import br.edu.ifgoias.academico.entities.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.edu.ifgoias.academico.entities.Curso;
-
 @Repository
-public interface CursoRepository extends JpaRepository<Curso, Integer>{
-
+public interface CursoRepository extends JpaRepository<Curso, Long> {
 }
